refactor: create an audio button component
